Merge the I/O refactoring branch. Every file format now has a corresponding RTLFile subclass, which contains a path or connection to data in that format. The original API remains intact, and is now a wrapper around the RTLFile API. This makes all file formats consistent with the TwoBit and BigWig handling, as well as the design of Rsamtools. It also allows other packages to extend rtracklayer, simply by subclassing RTLFile and implementing methods for import and export.
A major effort went into implementing tests for the import and export
methods, to ensure that the aforementioned refactoring did not break
anything. This led to the discovery of many bugs/annoyances that have
been fixed.

Also, every import method has gained a 'which' argument that will
return only the features that fall within 'which'. This takes
advantage of indexing like tabix, when available. This complements the
support for automatic indexing in many export methods.

This commit is a little premature: testing is not yet complete for the
WIG format.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/rtracklayer@64100 bc3139a8-67e5-0310-9ffc-ced21a209358